+++ source/qcommon/net_chan.cpp
@@ -18,7 +18,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 */
 
-#include "qcommon.h"
+#include "qcommon/qcommon.h"
 
 #if defined ( __MACOSX__ )
 #include <arpa/inet.h>
@@ -322,7 +322,7 @@ bool Netchan_TransmitNextFragment( netchan_t *chan ) {
 		fragmentLength = chan->unsentLength - chan->unsentFragmentStart;
 		last = true;
 	} else {
-		fragmentLength = ceil( ( chan->unsentLength - chan->unsentFragmentStart ) * 1.0 / ceil( ( chan->unsentLength - chan->unsentFragmentStart ) * 1.0 / FRAGMENT_SIZE ) );
+		fragmentLength = ceilf( ( chan->unsentLength - chan->unsentFragmentStart ) * 1.0 / ceilf( ( chan->unsentLength - chan->unsentFragmentStart ) * 1.0 / FRAGMENT_SIZE ) );
 		last = false;
 	}
